Week 1 + Week 2
As week 2's lab progress was built upon week 1's foundation. A single Repo was submitted.

Week 1: We began to build the foundation of the Lab's unity game. We learnt how to add modifiers to objects as well as C# scripts. Player controls were halfway implemented via keymapping, but the C# script was left unfinished due to the lack of time.

Week 2: We began to finish the player movement done in week 1. We also began to implement singleton patterns into the game. A additional point system was also implemented into the game.

Note: The modified version was combined to the original version